RAG_IMPLEMENTATION/requirements.txt
pdfplumber
faiss-cpu
sentence-transformers
transformers
torch
tqdm
